Authorization & Authentication on post & for its deletion

+++ controllers/api/v1/posts_api.js
@@ -26,31 +26,20 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
 
-        //if (post.user == req.user.id){
+        if (post.user == req.user.id){
             post.remove();
 
             await Comment.deleteMany({post : req.params.id});
 
             return res.json(200, {
                 message: 'post & associated comments deleted successfully!'
-            })
-
-            // if(req.xhr){
-            //     return res.status(200).json({
-            //         data:{
-            //             post_id : req.params.id
-            //         },
-            //         message: 'post deleted'
-            //     })
-            // }
-
-            //req.flash('error', 'Post & Comments Deleted')
-            //return res.redirect('back');
+            });
             
-        // }else{
-        //     req.flash('error', 'you cannot delete this post')
-        //     return res.redirect('back')
-        // }
+        }else{
+             return res.json(401, {
+                 message: 'You cannot delete it'
+             });
+         }
     }catch(err){
         //req.flash('error', err);
         console.log('****', err);
